geoweb/requirements.txt
Django
django-crispy-forms 
django-filter 
django-mptt 
django-mptt-tree-editor 
django-pagination-bootstrap 
django-registration-redux 
psycopg2 
setuptools 
six 
wheel
